+++ CHANGELOG.md
@@ -15,6 +15,17 @@ full changeset diff at the end of each section.
 Current Trunk
 -------------
 
+- Wast file parsing rules now don't allow a few invalid formats for typeuses
+  that were previously allowed. Typeuse entries should follow this format,
+  meaning they should have (type) -> (param) -> (result) order if more than one
+  of them exist.
+  ```
+  typeuse ::= (type index|name)+ |
+              (type index|name)+ (param ..)* (result ..)* |
+              (param ..)* (result ..)*
+  ```
+  Also, all (local) nodes in function definition should be after all typeuse
+  elements.
 - Removed APIs related to deprecated instruction names in Binaryen.js:
   - `get_local` / `getLocal`
   - `set_local` / `setLocal`